% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{build_download_url}
\alias{build_download_url}
\title{Build download URL for Iowa DOE data}
\usage{
build_download_url(media_id)
}
\arguments{
\item{media_id}{Media ID from get_media_id()}
}
\value{
Full download URL
}
\description{
Build download URL for Iowa DOE data
}
\keyword{internal}
