% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_historical_excel}
\alias{download_historical_excel}
\title{Download and read historical Iowa DOE Excel file}
\usage{
download_historical_excel(end_year)
}
\arguments{
\item{end_year}{School year end (1947-1991)}
}
\value{
Data frame with enrollment data for the specified year
}
\description{
Downloads a historical Excel file from Iowa DOE and extracts data for the
specified year. Historical files contain multiple years of data.
}
\keyword{internal}
