% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_available_years}
\alias{get_available_years}
\title{Get available years for Iowa enrollment data}
\usage{
get_available_years()
}
\value{
Integer vector of available years
}
\description{
Returns the range of years for which enrollment data is available.
Iowa has data from 1991-92 (end_year = 1992) through 2024-25 (end_year = 2025).
}
\examples{
get_available_years()
}
