% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_min_year}
\alias{get_min_year}
\title{Get the minimum available year}
\usage{
get_min_year()
}
\value{
Integer representing earliest available year
}
\description{
Iowa has enrollment data by grade, race, ethnicity, and gender from 1991-92.
}
\keyword{internal}
