% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_raw_cache_path}
\alias{get_raw_cache_path}
\title{Get path for raw file cache}
\usage{
get_raw_cache_path(end_year, level)
}
\arguments{
\item{end_year}{School year end}

\item{level}{Data level}
}
\value{
Path to raw cache file
}
\description{
Get path for raw file cache
}
\keyword{internal}
