% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{iowa_base_url}
\alias{iowa_base_url}
\title{Base URL for Iowa Department of Education data files}
\usage{
iowa_base_url()
}
\description{
Base URL for Iowa Department of Education data files
}
\keyword{internal}
