% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{is_historical_year}
\alias{is_historical_year}
\title{Check if year is in historical range}
\usage{
is_historical_year(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
TRUE if year is in historical range
}
\description{
Historical data (1947-1991) has a different format than modern data.
}
\keyword{internal}
