% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_historical_enr}
\alias{process_historical_enr}
\title{Process historical enrollment data}
\usage{
process_historical_enr(df, end_year)
}
\arguments{
\item{df}{Raw historical data frame}

\item{end_year}{School year end}
}
\value{
Processed historical data frame
}
\description{
Processes historical (1947-1991) enrollment data which has a simpler format
than modern data - grade-level enrollment only, no demographics or gender.
}
\keyword{internal}
